requirements.txt
aiogram==3.22.0
aiohttp>=3.9
aiosqlite>=0.19
python-dotenv>=1.0
colorama>=0.4.6